backend/requirements-prod.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx[http2]==0.25.0
python-multipart==0.0.6
anyio==3.7.1
sniffio==1.3.0
starlette==0.27.0
